update flatten action outputs to consider OutputSymlinks
* Fix readStreamedRetried retry in bytestream (#619)

Fix readStreamedRetried in rbe bytestream

Transient errors are retried by `readStreamedRetried` and it continues reading from how many bytes were already read. When an error happens in the stream call (for example, `DeadlineExceeded`), we were returning that 0 bytes were read.

Since `DeadlineExceeded` is transient, the retry happens and we start from byte 0 again, leading to a lot of duplication.

Note that the other errors return a `fmt.Errorf`, which is not a transient error and won't be retried.

The current behavior seems to be only working to retry at the start of the stream, when no bytes were there yet.

* Fix bug in handling of expiry buffer. (#620)

* update flatten action outputs to consider OutputSymlinks

OutputSymlinks is the non-deprecated way to provide symlinks, but not handled in the FlattenActionOutputs.

* Update test to account outputsymlink

---------

Co-authored-by: ramymedhat <abdelaal@google.com>